feat: Replace render-demo.js with improved render-pdf.js script (#145)
- Remove render-demo.js
- Add new render-pdf.js with automatic output naming
- Update README.md documentation
- Script now auto-detects schema type and version
- Saves PDF in same directory as input by default
- Supports optional --output parameter for custom directory
